requirements.txt
lorem>=0.1.1
certifi>=2025.8.3
charset-normalizer>=3.4.3
idna>=3.10
iniconfig>=2.1.0
packaging>=25.0
pillow>=11.3.0
pluggy>=1.6.0
Pygments>=2.19.2
pytest>=8.4.2
requests>=2.32.5
urllib3>=2.5.0
